Added files to project
models folder was added with 1 file. The category class is created
inside the folder.

dummy data was added.

categories screen was added with a grid. A grid is different from a
list. A grid can have items horizontally as well as vertically as
opposed to a list where items are displayed vertically.